# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Engram API
  description: |
    Lore persistence and synchronization service for AI agent memory.

    Engram provides a central repository for experiential knowledge (lore) that
    AI agents accumulate during development sessions. Features include semantic
    deduplication, confidence scoring, feedback integration, and efficient sync.

    ## Multi-Store Support (v1.1)

    Engram supports isolated stores for multi-project deployments. Each store
    maintains its own lore collection with independent snapshots and statistics.

    - **Default store**: Auto-created, used by legacy `/lore/*` routes
    - **Named stores**: Created via `POST /stores`, accessed via `/stores/{store_id}/lore/*`
    - **Backward compatible**: Existing clients work without changes
  version: 1.1.0
  contact:
    name: Engram Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://your-engram-instance.example.com/api/v1
    description: Production (replace with your deployment URL)
...
paths:
...
  /stats:
    get:
      summary: Get extended system statistics
      description: |
        Returns comprehensive system metrics including lore counts, embedding pipeline health,
        category distribution, and quality metrics. This endpoint is public (no authentication
        required) to enable monitoring tool integration.
      operationId: getStats
      tags:
        - Monitoring
      security: []
      responses:
        '200':
          description: Extended statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedStats'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # === Store Management Endpoints ===
  /stores:
    get:
      tags: [Stores]
      summary: List all stores
      description: |
        Returns a list of all stores with summary information including
        record counts and storage sizes. Stores are sorted alphabetically by ID.
      operationId: listStores
      responses:
        '200':
          description: List of stores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListStoresResponse'
              example:
                stores:
                  - id: "default"
                    record_count: 1523
                    last_accessed: "2026-01-31T10:30:00Z"
                    size_bytes: 4194304
                    description: "Default store (auto-created)"
                  - id: "neuralmux/engram"
                    record_count: 847
                    last_accessed: "2026-01-31T09:15:00Z"
                    size_bytes: 2097152
                    description: "Engram project lore"
                total: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Multi-store support not configured
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    post:
      tags: [Stores]
      summary: Create a new store
      description: |
        Creates a new isolated store for lore entries. The default store is
        auto-created on first access and does not need explicit creation.

        **Store ID Format:**
        - Lowercase alphanumeric characters and hyphens only
        - Path segments separated by `/` (max 4 levels)
        - Maximum 128 characters total
        - Must start and end with alphanumeric (no leading/trailing hyphens)
        - Examples: `myproject`, `org/team/project`, `neuralmux/engram`
      operationId: createStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreRequest'
            example:
              store_id: "neuralmux/recall"
              description: "Recall client lore"
      responses:
        '201':
          description: Store created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateStoreResponse'
              example:
                id: "neuralmux/recall"
                created: "2026-01-31T11:00:00Z"
                description: "Recall client lore"
        '400':
          description: Invalid store ID format
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://engram.dev/errors/bad-request"
                title: "Bad Request"
                status: 400
                detail: "invalid store ID: must be lowercase alphanumeric with hyphens"
                instance: "/api/v1/stores"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Store already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://engram.dev/errors/conflict"
                title: "Conflict"
                status: 409
                detail: "Store already exists: neuralmux/recall"
                instance: "/api/v1/stores"
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Multi-store support not configured
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /stores/{store_id}:
    get:
      tags: [Stores]
      summary: Get store information
      description: |
        Returns detailed information about a specific store including
        extended statistics (lore counts, embedding stats, quality metrics).
      operationId: getStoreInfo
      parameters:
        - $ref: '#/components/parameters/StoreIdPath'
      responses:
        '200':
          description: Store information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreInfoResponse'
              example:
                id: "neuralmux/engram"
                created: "2026-01-15T08:00:00Z"
                last_accessed: "2026-01-31T10:30:00Z"
                description: "Engram project lore"
                size_bytes: 2097152
                stats:
                  total_lore: 850
                  active_lore: 847
                  deleted_lore: 3
                  embedding_stats:
                    complete: 840
                    pending: 7
                    failed: 0
                  category_stats:
                    ARCHITECTURAL_DECISION: 125
                    PATTERN_OUTCOME: 89
                  quality_stats:
                    average_confidence: 0.72
                    validated_count: 156
                    high_confidence_count: 312
                    low_confidence_count: 45
                  unique_source_count: 8
                  stats_as_of: "2026-01-31T10:30:00Z"
        '400':
          description: Invalid store ID format
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Store not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Multi-store support not configured
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      tags: [Stores]
      summary: Delete a store
      description: |
        Permanently deletes a store and all its lore data.

        **Safeguards:**
        - Requires `confirm=true` query parameter
        - The `default` store cannot be deleted (returns 403)

        **Warning:** This operation is irreversible.
      operationId: deleteStore
      parameters:
        - $ref: '#/components/parameters/StoreIdPath'
        - name: confirm
          in: query
          required: true
          schema:
            type: string
            enum: ["true"]
          description: Must be `true` to confirm deletion
      responses:
        '204':
          description: Store deleted successfully
        '400':
          description: Missing confirm=true parameter or invalid store ID
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://engram.dev/errors/bad-request"
                title: "Bad Request"
                status: 400
                detail: "Delete requires confirm=true query parameter"
                instance: "/api/v1/stores/neuralmux%2Fengram"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Cannot delete the default store
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://engram.dev/errors/forbidden"
                title: "Forbidden"
                status: 403
                detail: "Cannot delete the default store"
                instance: "/api/v1/stores/default"
        '404':
          description: Store not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Multi-store support not configured
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication via Bearer token.
        Include header: `Authorization: Bearer <your-api-key>`
  parameters:
    StoreIdPath:
      name: store_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/StoreID'
      description: |
        Store identifier. Must be URL-encoded if it contains forward slashes.
        Example: `neuralmux%2Fengram` for `neuralmux/engram`
  schemas:
...
    ListStoresResponse:
      type: object
      description: Response for listing all stores
      required:
        - stores
        - total
      properties:
        stores:
          type: array
          items:
            $ref: '#/components/schemas/StoreListItem'
          description: List of stores sorted alphabetically by ID
        total:
          type: integer
          description: Total number of stores
          example: 2
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the error type
          example: "https://engram.dev/errors/validation-error"
        title:
          type: string
          description: Human-readable error summary
          example: "Validation Error"
        status:
          type: integer
          description: HTTP status code
          example: 422
        detail:
          type: string
          description: Human-readable explanation
          example: "Request contains invalid fields"
        instance:
          type: string
          description: URI of the request that caused the error
          example: "/api/v1/lore"
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://engram.dev/errors/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "Missing or invalid Authorization header"
            instance: "/api/v1/lore"
security:
  - bearerAuth: []